+++ esp_cam_sensor/sensors/ov5647/ov5647.c
@@ -30,7 +30,6 @@
 #define OV5647_SUPPORT_NUM CONFIG_CAMERA_OV5647_MAX_SUPPORT
 
 static const char *TAG = "ov5647";
-static uint8_t s_ov5647_index;
 
 static const esp_cam_sensor_isp_info_t ov5647_isp_info[] = {
     {
@@ -456,7 +455,6 @@ static esp_err_t ov5647_delete(esp_cam_sensor_device_t *dev)
     if (dev) {
         free(dev);
         dev = NULL;
-        s_ov5647_index--;
     }
 
     return ESP_OK;
@@ -480,11 +478,6 @@ esp_cam_sensor_device_t *ov5647_detect(esp_cam_sensor_config_t *config)
     esp_cam_sensor_device_t *dev = NULL;
 
     if (config == NULL) {
-        return NULL;
-    }
-
-    if (s_ov5647_index >= OV5647_SUPPORT_NUM) {
-        ESP_LOGE(TAG, "Only support max %d cameras", OV5647_SUPPORT_NUM);
         return NULL;
     }
 
@@ -520,9 +513,7 @@ esp_cam_sensor_device_t *ov5647_detect(esp_cam_sensor_config_t *config)
         ESP_LOGE(TAG, "Camera sensor is not OV5647, PID=0x%x", dev->id.pid);
         goto err_free_handler;
     }
-    ESP_LOGI(TAG, "Detected Camera sensor PID=0x%x with index %d", dev->id.pid, s_ov5647_index);
-
-    s_ov5647_index++;
+    ESP_LOGI(TAG, "Detected Camera sensor PID=0x%x", dev->id.pid);
 
     return dev;